jlr-api-function/requirements.txt
jlrpy
google-cloud-storage